+++ plan.md
@@ -1,4 +1,4 @@
 ✅ `src/shared/ui/FollowButton` を `src/features/follow/ui` へ移動する。
 
 ✅ `src/shared/ui/UsersList` を `src/widgets/user-list` へ移動する
-- [ ] `
+✅ `src/shared/ui/Post/Post.tsx` を `src/entities/post` へ移動する
