+++ src/entities/post/index.ts
@@ -1,0 +1,1 @@
+export { Post } from "./ui";
