+++ src/pages/posts/ui/PostDetailPageClient.tsx
@@ -1,8 +1,9 @@
 "use client";
 
+import { Post } from "@/entities/post";
 import type { PostType, User } from "@/shared/entities";
 import { sideBarWidthBase } from "@/shared/lib/const";
-import { Post, Sidebar } from "@/shared/ui";
+import { Sidebar } from "@/shared/ui";
 import { Box, Flex, Space } from "@mantine/core";
 import { useRouter } from "next/navigation";
