+++ src/pages/top/ui/TopPageClient.tsx
@@ -2,7 +2,8 @@
 
 import type { PostType, User } from "@/shared/entities";
 import { sideBarWidthBase } from "@/shared/lib/const";
-import { NewPostForm, PostsList, Sidebar } from "@/shared/ui";
+import { NewPostForm, Sidebar } from "@/shared/ui";
+import { PostsList } from "@/widgets/posts-list";
 import { Box, Flex } from "@mantine/core";
 
 interface TopPageClientProps {
