+++ src/shared/ui/index.ts
@@ -1,7 +1,9 @@
 export { Sidebar } from "./Sidebar/Sidebar";
 export { ImageUpload } from "./ImageUpload/ImageUpload";
-export { Post } from "./Post/Post";
 export { EditProfileModal } from "./Modal/EditProfileModal/EditProfileModal";
-export { PostsList } from "./PostsList/PostsList";
 export { NewPostForm } from "./NewPostForm/NewPostForm";
 export { CenteredLoader } from "./CenteredLoader/CenteredLoader";
+export { ContentPart, parseContent } from "./ContentPart/ContentPart";
+export { DateFormat } from "./DateFormat/DateFormat";
+export { PostActionMenu } from "./PostActionMenu/PostActionMenu";
+export { PostActionsButtonGroup } from "./PostActionsButtonGroup/PostActionsButtonGroup";
