+++ src/widgets/posts-list/index.ts
@@ -1,0 +1,1 @@
+export { PostsList } from "./ui";
